refactor(#155 PR 7/10): extract sim/topology.js from game.js
Pure code movement, zero behavior change. game.js: 2318 -> 2062 lines
(4076 at the start of the split).

- src/sim/topology.js (316 lines): the build/wire/demolish cluster —
  snapToGrid, createService, restoreService, createConnection (with the
  connection-validity table and the #192 reverse-edge guard),
  deleteConnection, getConnectionAtPoint, deleteObject,
  updateConnectionsForNode, clearAllServices. All byte-verbatim
  (diff-verified against HEAD line ranges).
- save-load.js and campaign-ui.js repointed to import topology fns from
  the new module; game.js exports its raycasting singletons for the
  runtime-only cycle (established pattern).
- window.clearAllServices stays in game.js's boundary block.

Verified: lint 0, 87/87 tests, 29-module graph links, in-browser —
full lifecycle through the new module (place 4 services, wire 4
connections, 40/40 requests processed, delete connection, delete
service, clear all), the #191/#192 rules hold after the move (alb->sqs
links, reverse edge blocked), campaign level 2 still prebuilds, zero
console errors.
